requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
protobuf==4.25.1
scikit-learn==1.3.2
numpy==1.26.2
